train/requirements.txt
dlib==19.21.0
numpy==1.19.1
torchvision==0.4.2
opencv_python==4.4.0.42
scipy==1.2.1
torch==1.3.1
Pillow==8.0.1
tensorboardX==2.1
